+++ 4chan Scraper.py
@@ -1,207 +1,147 @@
-##################################
-#####BATEMAN'S 4CHAN SCRAPER######
-##################################
-##github.com/SelfAdjointOperator##
-##################################
+
+"""https://github.com/SelfAdjointOperator/4chan-Scraper"""
 
 import urllib.request       #   getting files from web
 import json                 #   config file and api pages jsons to and from dictionary
-import os                   #   managing folders and update files
+import os                   #   managing folders
 import threading            #   multiple simultaneous downloads
-from sys import stdout      #   for progress bar
 from time import sleep,time #   sleep if 4plebs search cooldown reached, restart delay
-# from hashlib import md5   #   hashing already scraped files if number not in active : currently not in use
 
-version = '2.1.4'
+# SAO Suite imports
+from saosuite import saotitle
+from saosuite import saostatusmsgs
+from saosuite import saoconfigmanager
+from saosuite import saovcs
+from saosuite import saomd5
+
+version = '3.0.0'
 boxestocheckfor = {"4chan":["name","sub","com","filename"],"4plebs":["username","subject","text","filename"]}
-plebboards = ['adv','f','hr','o','pol','s4s','sp','tg','trv','tv','x']
+plebBoards = ['adv','f','hr','o','pol','s4s','sp','tg','trv','tv','x']
 plebsHTTPHeader = 'Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US; rv:1.9.0.7) Gecko/2009021910 Firefox/3.0.7'
 num_download_threads = 4
 
 ################################################################################
 
-def new_config():
-    global version
-    return {"versioncreated":version, "boards":{}}
-
-################################################################################
-
-def new_board():
-    return {"keywords":[], "blacklist":[], "requests":[], "active":[], "doneops":[]}
-    # active/requests example [opno,keyword,[scrapednos]]
-
-################################################################################
-
-def possible_new_board(boardcode):
-    global configjson
-    if not boardcode in configjson["boards"]:
-        configjson["boards"][boardcode] = new_board()
-
-################################################################################
-
-def saveconfig():
-    with open('scraperconfig.json','w') as configjson_file:
-        configjson_file.write(json.dumps(configjson))
-
-################################################################################
-
 def scrape():
-    nonemptyBoards_requests = [b for b in configjson["boards"] if configjson["boards"][b]["requests"]]
-    if not nonemptyBoards_requests:
+    cm.disableAutosave()
+    # do special requests
+    nonemptyBoards_special = [board for board in cm.valueGet("downloaded") if cm.tpt_getTasksInTier("downloaded/{}".format(board),"special")]
+    if not nonemptyBoards_special:
         print("Currently no special requests")
     else:
         print("~Doing special requests~")
-        requestsToDo = []
-        for board in nonemptyBoards_requests:
-            for req in configjson["boards"][board]["requests"]:
-                if req[0] in configjson["boards"][board]["doneops"]:
-                    print("Already scraped /{}/:{}:{}".format(board,str(req[0]),req[1]))
-                else:
-                    requestsToDo.append([board,req[0],req[1],req[2],len(board)+len(str(req[0]))+len(req[1])])
-            configjson["boards"][board]["requests"] = []
-        if requestsToDo:
-            maxpad = max([rtd[4] for rtd in requestsToDo])
-            for rtd in requestsToDo:
-                rtd[4] = maxpad - rtd[4]
-                result = scrapethread(*rtd)
+        paddingLength = 0
+        for board in nonemptyBoards_special:
+            for task in cm.tpt_pruneTasks("downloaded/{}".format(board),tiers=["special"],idnos_done=True)["special"]:
+                print("Already scraped /{}/:{}:{}".format(board,str(task[0]),task[1]))
+            try:
+                paddingLength = max(paddingLength,max([len(task[1]) for task in cm.tpt_getTasksInTier("downloaded/{}".format(board),"special")]))
+            except: # if no tasks left in tier
+                pass
+        for board in nonemptyBoards_special:
+            for task in [t for t in cm.tpt_getTasksInTier("downloaded/{}".format(board),"special")]:
+                result = scrapeThread(board,*task,paddingLength)
                 if result[0] == 'keep':
-                    configjson["boards"][rtd[0]]["requests"].append([rtd[1],rtd[2],result[1]])
+                    cm.tpt_updateTaskByIdno("downloaded/{}".format(board),task[0],result[1])
                 else:
-                    configjson["boards"][rtd[0]]["doneops"].insert(0,rtd[1])
-
+                    cm.tpt_finishTaskByIdno("downloaded/{}".format(board),task[0])
     print()
-    nonemptyBoards_keywords = [b for b in configjson["boards"] if configjson["boards"][b]["keywords"]]
-    if not nonemptyBoards_keywords:
+
+    # do normal scraping
+    if not (nonemptyBoards_keywords := [board for board in cm.valueGet("downloaded") if cm.tpt_getkeywords_wl("downloaded/{}".format(board))]):
         print("Currently not scraping any boards\n")
     else:
         for board in nonemptyBoards_keywords:
-            [all_doneops,all_active,blacklist_return] = scrapeboard(board,configjson["boards"][board]["keywords"],configjson["boards"][board]["blacklist"],[t[0] for t in configjson["boards"][board]["requests"]],configjson["boards"][board]["active"],configjson["boards"][board]["doneops"])
-            configjson["boards"][board]["doneops"] = all_doneops
-            configjson["boards"][board]["active"] = all_active
-            configjson["boards"][board]["blacklist"] = blacklist_return
-            print()
+            keywords_wl = cm.tpt_getkeywords_wl("downloaded/{}".format(board))
+            idnos_bl = cm.tpt_getidnos_bl("downloaded/{}".format(board))
+            idnos_done = cm.tpt_getidnos_done("downloaded/{}".format(board))
+
+            #Board Catalog JSON
+            try:
+                print("~Getting JSON for catalog of /{}/~".format(board))
+                catalogjson_url = ("https://a.4cdn.org/{}/catalog.json".format(board))
+                catalogjson_file = urllib.request.urlopen(catalogjson_url)
+                catalogjson = json.load(catalogjson_file)
+                blacklist_expired = [t for t in idnos_bl]
+                #Search ops not considered already
+                for page in catalogjson:
+                    for threadop in page["threads"]:
+                        opno = threadop["no"]
+                        if opno in idnos_bl:
+                            blacklist_expired.remove(opno)
+                        if opno in idnos_done:
+                            continue
+                        boxbreak = False
+                        boxestocheck=[box for box in boxestocheckfor["4chan"] if box in threadop]
+                        for boxtocheck in boxestocheck:
+                            for keyword in keywords_wl:
+                                if keyword in threadop[boxtocheck].lower():
+                                    cm.tpt_promoteTaskToByIdno("downloaded/{}".format(board),opno,keyword=keyword,promotionTier="normal") # manage folders here
+                                    boxbreak = True
+                                    break
+                            if boxbreak:
+                                break
+                for task in cm.tpt_pruneTasks("downloaded/{}".format(board),tiers=["normal"],keywords_wl=True,idnos_bl=True,idnos_done=True)["normal"]:
+                    cm.ffm_rmIfEmptyTree("downloaded/{}/{} {}".format(board,str(task[0]),task[1]))
+                tasksToScrape = cm.tpt_getTasksInTier("downloaded/{}".format(board),"normal")[:]
+                paddingLength = max([len(task[1]) for task in tasksToScrape]) if tasksToScrape else 0
+                for task in tasksToScrape:
+                    if (result := scrapeThread(board,*task,paddingLength))[0] == 'keep':
+                        cm.tpt_updateTaskByIdno("downloaded/{}".format(board),task[0],result[1])
+                    else:
+                        cm.tpt_finishTaskByIdno("downloaded/{}".format(board),task[0])
+                cm.tpt_idnos_blRemove("downloaded/{}".format(board),blacklist_expired)
+                print()
+            except:
+                print("Error: Cannot load catalog for /{}/".format(board))
+
+    for board in cm.valueGet("downloaded"):
+        cm.ffm_rmIfEmptyTree("downloaded/{}".format(board))
+
     print("~Updating config~")
-    saveconfig()
+    cm.enableAutosave()
     print("~Config updated~")
     print("~Done scraping~")
 
 ################################################################################
 
-def scrapeboard(boardcode,keywords,blacklist,requestopnos,active,doneops):
-    global boxestocheckfor
-    alreadyConsidered_opnos = blacklist + requestopnos + doneops
-
-    #Check if current active are still what we want
-    threadstoscrape = [t for t in active if not t[0] in alreadyConsidered_opnos and t[1] in keywords]
-    alreadyConsidered_opnos += [t[0] for t in threadstoscrape]
-    formerRequests = [t for t in active if t[1][:8] == '_OLDREQ_']
-    active_return = []
-
-    #Board Catalog JSON
-    try:
-        print("~Getting JSON for catalog of /{}/~".format(boardcode))
-        catalogjson_url = ("https://a.4cdn.org/{}/catalog.json".format(boardcode))
-        catalogjson_file = urllib.request.urlopen(catalogjson_url)
-        catalogjson = json.load(catalogjson_file)
-        blacklist_return = []
-        #Search ops not considered already
-        for page in catalogjson:
-            for threadop in page["threads"]:
-                opno = threadop["no"]
-                if opno in blacklist:
-                    blacklist_return.append(opno)
-                if not opno in alreadyConsidered_opnos:
-                    boxbreak=0
-                    boxestocheck=[b for b in boxestocheckfor["4chan"] if b in threadop]
-                    for boxtocheck in boxestocheck:
-                        for keyword in keywords:
-                            if keyword in threadop[boxtocheck].lower():
-                                try: #check if former request with scraped nos already
-                                    formerRequest = [t for t in formerRequests if t[0] == opno][0]
-                                    formerRequest_nos = formerRequest[2]
-                                    try:
-                                        folder_old = "{}\\{} {}".format(boardcode,str(opno),formerRequest[1][8:])
-                                        folder_new = "{}\\{} {}".format(boardcode,str(opno),keyword)
-                                        os.rename(folder_old,folder_new)
-                                    except:
-                                        pass
-                                except:
-                                    formerRequest_nos = []
-                                threadstoscrape.append([opno,keyword,formerRequest_nos])
-                                boxbreak=1
-                                break
-                        if boxbreak==1:
-                            break
-    except:
-        print("Error: Cannot load catalog for /{}/".format(boardcode))
-        blacklist_return = blacklist
-
-    if threadstoscrape:
-        #Compute padding for progress bar placement:
-        maxsize = max([len(str(t[0]))+len(t[1]) for t in threadstoscrape])
-        threadstoscrape_padded = [t+[maxsize-(len(str(t[0]))+len(t[1]))] for t in threadstoscrape]
-        #Actually do the scraping now
-        for ttsp in threadstoscrape_padded:
-            result = scrapethread(boardcode,*ttsp)
-            if result[0] == 'keep':
-                active_return.append([ttsp[0],ttsp[1],result[1]])
-            else:
-                doneops.insert(0,ttsp[0])
-
-    return [doneops,active_return,blacklist_return]
-
-################################################################################
-
-def scrapethread(boardcode,threadopno,keyword,scrapednos,padding):
+def scrapeThread(boardcode,threadopno,keyword,scrapednos,padding):
     global lock
-    filelist = getfilelist(boardcode,threadopno,keyword,'4chan')
+    filelist = getFileList(boardcode,threadopno,keyword,'4chan')
     filestart = 0
     if filelist[0] in ['try_4plebs']:
-        filelist = getfilelist(boardcode,threadopno,keyword,'4plebs')
+        filelist = getFileList(boardcode,threadopno,keyword,'4plebs')
         filestart = 1
     if filelist[0] in ['keep','delete']:
         return [filelist[0],scrapednos]
     #otherwise 'now_scrape'
     impostslist = filelist[1]
 
-    #Try to create folder
-    threadaddress=("{}\\{} {}".format(boardcode,str(threadopno),keyword))
-    try:
-        os.makedirs(threadaddress,exist_ok=True)
-    except:
-        print("Error: failed to create folder '{}'".format(threadaddress))
-        return ['keep',scrapednos]
+    threadaddress=("downloaded/{}/{} {}".format(boardcode,str(threadopno),keyword))
+    cm.ffm_makedirs("{}/thumbs".format(threadaddress))
 
     #Scrape files
-    progressmsg.progmsg(msg="Scraping /{}/:{}:{} {}".format(boardcode,str(threadopno),keyword,' '*padding),of=len(impostslist))
+    pm.progressmsg(msg="Scraping /{}/:{}:{} ".format(boardcode,str(threadopno),keyword.ljust(padding)),of=len(impostslist))
     keepflag = 0
 
     postbuffers = [[] for i in range(num_download_threads)]
-    def scrapefile_download_thread(dtid):
+    def scrapeFile_download_thread(dtid):
         nonlocal keepflag, postbuffers
         while True:
             with lock:
                 try:
                     postbuffers[dtid] = impostslist.pop(0)
                     if postbuffers[dtid]["no"] in scrapednos:
-                        progressmsg.tick()
+                        pm.tick()
                         continue
-                except:
+                except IndexError:
                     return
             for modus in ['4chan','4plebs','4plebsthumbs'][filestart:]:
-                if modus == '4plebsthumbs':
-                    with lock:
-                        try:
-                            os.makedirs('{}\\thumbs'.format(threadaddress),exist_ok=True)
-                        except:
-                            progressmsg.progmsg(msg="Error: failed to create folder \'{}\\thumbs\' ".format(threadaddress))
-                            keepflag = 1
-                            break
-                result = scrapefile(threadaddress,postbuffers[dtid],modus,boardcode,threadopno,keyword)
+                result = scrapeFile(threadaddress,postbuffers[dtid],modus,boardcode,threadopno,keyword)
                 with lock:
                     if result == 'success':
                         scrapednos.append(postbuffers[dtid]["no"])
-                        progressmsg.tick()
+                        pm.tick()
                         break
                     elif result == 'keep':
                         keepflag = 1
@@ -209,24 +149,15 @@ def scrapethread(boardcode,threadopno,keyword,scrapednos,padding):
                     elif result == 'try_next_modus':
                         continue
 
-    download_threads = [threading.Thread(target=scrapefile_download_thread,args=[i]) for i in range(num_download_threads)]
+    download_threads = [threading.Thread(target=scrapeFile_download_thread,args=[i]) for i in range(num_download_threads)]
     for t in download_threads:
         t.start()
     for t in download_threads:
         t.join()
-    progressmsg.finish()
+    pm.finish()
 
-    #Delete empty folder (or / and thumbs subfolder)
-    if 'thumbs' in os.listdir(threadaddress) and not [f for f in os.listdir('{}\\thumbs'.format(threadaddress))]:
-        try:
-            os.rmdir('{}\\thumbs'.format(threadaddress))
-        except:
-            print("Error: Could not delete folder '{}\\thumbs'".format(threadaddress))
-    if not [f for f in os.listdir(threadaddress)]:
-        try:
-            os.rmdir(threadaddress)
-        except:
-            print("Error: Could not delete folder '{}'".format(threadaddress))
+    cm.ffm_rmIfEmptyTree(threadaddress+"/thumbs")
+    cm.ffm_rmIfEmptyTree(threadaddress)
 
     if keepflag == 0 and (filestart!=0 or filelist[2] == True):
         return ['delete']
@@ -235,9 +166,7 @@ def scrapethread(boardcode,threadopno,keyword,scrapednos,padding):
 
 ################################################################################
 
-def getfilelist(boardcode,threadopno,keyword,modus,silent=False):
-    global plebboards,plebsHTTPHeader
-
+def getFileList(boardcode,threadopno,keyword,modus):
     def gfl_error(num):
         gfl_errors = [
             "Thread /{}/:{}:{} not found on 4chan, trying 4plebs",
@@ -245,21 +174,18 @@ def getfilelist(boardcode,threadopno,keyword,modus,silent=False):
             "Error: Cannot load 4chan thread /{}/:{}:{}",
             "Thread /{}/:{}:{} not found on 4plebs",
             "Error: Cannot load 4plebs thread /{}/:{}:{}"]
-        if silent is False:
-            print(gfl_errors[num].format(boardcode,str(threadopno),keyword))
+        print(gfl_errors[num].format(boardcode,str(threadopno),keyword))
 
     if modus == '4chan':
         try:
             threadjson_url = 'https://a.4cdn.org/{}/thread/{}.json'.format(boardcode,str(threadopno))
             threadjson_file = urllib.request.urlopen(threadjson_url)
             threadjson = json.load(threadjson_file)
-            impostslist = [{"no":post['no'],"tim":post['tim'],"ext":post['ext']} for post in threadjson["posts"] if "tim" in post]
+            impostslist = [{"no":post['no'],"tim":post['tim'],"ext":post['ext'],"md564":post['md5']} for post in threadjson["posts"] if "tim" in post]
             return ['now_scrape',impostslist,'archived' in threadjson["posts"][0]]
         except Exception as e:
-            #Thread error:
             if hasattr(e,'code') and e.code == 404: # pylint: disable=E1101
-                if boardcode in plebboards:
-                    #If 404 error and plebboard then try to get thread JSON from 4plebs
+                if boardcode in plebBoards:
                     gfl_error(0)
                     return ['try_4plebs']
                 else:
@@ -281,11 +207,11 @@ def getfilelist(boardcode,threadopno,keyword,modus,silent=False):
                     raise Exception
             impostslist = []
             if "op" in threadjson[str(threadopno)] and threadjson[str(threadopno)]["op"]["media"] != None:
-                impostslist.append({"no":int(threadjson[str(threadopno)]["op"]["num"]),"tim":os.path.splitext(threadjson[str(threadopno)]["op"]["media"]["media"])[0],"ext":os.path.splitext(threadjson[str(threadopno)]["op"]["media"]["media"])[1]})
+                impostslist.append({"no":int(threadjson[str(threadopno)]["op"]["num"]),"tim":os.path.splitext(threadjson[str(threadopno)]["op"]["media"]["media"])[0],"ext":os.path.splitext(threadjson[str(threadopno)]["op"]["media"]["media"])[1],"md564":threadjson[str(threadopno)]["op"]["media"]["media_hash"]})
             if "posts" in threadjson[str(threadopno)]:
                 for postvalue in threadjson[str(threadopno)]["posts"].values():
                     if postvalue["media"] != None:
-                        impostslist.append({"no":int(postvalue["num"]),"tim":os.path.splitext(postvalue["media"]["media"])[0],"ext":os.path.splitext(postvalue["media"]["media"])[1]})
+                        impostslist.append({"no":int(postvalue["num"]),"tim":os.path.splitext(postvalue["media"]["media"])[0],"ext":os.path.splitext(postvalue["media"]["media"])[1],"md564":postvalue["media"]["media_hash"]})
             return ['now_scrape',impostslist]
         except Exception as e:
             if hasattr(e,'code') and e.code in [404,'404']: # pylint: disable=E1101
@@ -297,39 +223,44 @@ def getfilelist(boardcode,threadopno,keyword,modus,silent=False):
 
 ################################################################################
 
-def scrapefile(threadaddress,post,modus,boardcode,threadopno,keyword):
-    global plebboards
-
+def scrapeFile(threadaddress,post,modus,boardcode,threadopno,keyword):
     def sf_error(num):
         global lock
         sf_errors = [
-            "File /{}/:{}:{}:{} already exists; possible duplicate scraped ",
+            "File /{}/:{}:{}:{} already exists with different MD5 checksum; possible duplicate scraped ",
             "File /{}/:{}:{}:{} not found on 4chan, scraping 4plebs file ",
             "File /{}/:{}:{}:{} not found on 4chan and not on 4plebs ",
             "Error: Cannot load 4chan file /{}/:{}:{}:{} ",
-            "File /{}/:{}:{}:{} already exists; possible duplicate scraped ",
+            "File /{}/:{}:{}:{} already exists with different MD5 checksum; possible duplicate scraped ",
             "File /{}/:{}:{}:{} not found on 4plebs, scraping 4plebs thumbnail ",
             "Error: Cannot load 4plebs file /{}/:{}:{}:{} ",
-            "File /{}/:{}:{}:{}(thumb) already exists; possible duplicate scraped ",
+            "File /{}/:{}:{}:{}(thumb) already exists with different MD5 checksum; possible duplicate scraped ",
             "File /{}/:{}:{}:{}(thumb) not found on 4plebs ",
-            "Error: Cannot load 4plebs file /{}/:{}:{}:{}(thumb) "]
+            "Error: Cannot load 4plebs file /{}/:{}:{}:{}(thumb) ",
+            "File /{}/:{}:{}:{} already exists with same MD5 checksum; not scraping again ",
+            "File /{}/:{}:{}:{} already exists with same MD5 checksum; not scraping again ",
+            "File /{}/:{}:{}:{}(thumb) already exists with same MD5 checksum; not scraping again "]
         with lock:
-            progressmsg.progmsg(msg=sf_errors[num].format(boardcode,threadopno,keyword,str(post["no"])))
+            pm.progressmsg(msg=sf_errors[num].format(boardcode,threadopno,keyword,str(post["no"])))
 
     if modus == '4chan':
         try:
-            imgaddress = "{}\\{}{}".format(threadaddress,str(post["no"]),post["ext"])
+            imgaddress = "{}/{}{}".format(threadaddress,str(post["no"]),post["ext"])
             if os.path.exists(imgaddress):
-                sf_error(0)
-                rn_name,rn_ext = os.path.splitext(imgaddress)
-                os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
+                if saomd5.isHashHex(imgaddress,saomd5.base64ToHex(post["md564"])):
+                    sf_error(10)
+                    return 'success'
+                else:
+                    sf_error(0)
+                    rn_name,rn_ext = os.path.splitext(imgaddress)
+                    os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
             imgdomain = 'https://i.4cdn.org/'
             imgurl = "{}{}/{}{}".format(imgdomain,boardcode,str(post["tim"]),post["ext"])
             urllib.request.urlretrieve(imgurl,imgaddress)
             return 'success'
         except Exception as e:
             if hasattr(e,'code') and e.code == 404: # pylint: disable=E1101
-                if boardcode in plebboards:
+                if boardcode in plebBoards:
                     sf_error(1)
                     return 'try_next_modus'
                 else:
@@ -341,11 +272,15 @@ def scrapefile(threadaddress,post,modus,boardcode,threadopno,keyword):
 
     elif modus == '4plebs':
         try:
-            imgaddress = "{}\\{}{}".format(threadaddress,str(post["no"]),post["ext"])
+            imgaddress = "{}/{}{}".format(threadaddress,str(post["no"]),post["ext"])
             if os.path.exists(imgaddress):
-                sf_error(4)
-                rn_name,rn_ext = os.path.splitext(imgaddress)
-                os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
+                if saomd5.isHashHex(imgaddress,saomd5.base64ToHex(post["md564"])):
+                    sf_error(11)
+                    return 'success'
+                else:
+                    sf_error(4)
+                    rn_name,rn_ext = os.path.splitext(imgaddress)
+                    os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
             imgdomain = 'https://i.4pcdn.org/'
             imgurl = "{}{}/{}{}".format(imgdomain,boardcode,str(post["tim"]),post["ext"])
             urllib.request.urlretrieve(imgurl,imgaddress)
@@ -360,12 +295,16 @@ def scrapefile(threadaddress,post,modus,boardcode,threadopno,keyword):
 
     elif modus == '4plebsthumbs':
         try:
-            threadaddress = '{}\\thumbs'.format(threadaddress)
-            imgaddress = "{}\\{}.jpg".format(threadaddress,str(post["no"]))
+            threadaddress = '{}/thumbs'.format(threadaddress)
+            imgaddress = "{}/{}.jpg".format(threadaddress,str(post["no"]))
             if os.path.exists(imgaddress):
-                sf_error(7)
-                rn_name,rn_ext = os.path.splitext(imgaddress)
-                os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
+                if saomd5.isHashHex(imgaddress,saomd5.base64ToHex(post["md564"])):
+                    sf_error(12)
+                    return 'success'
+                else:
+                    sf_error(7)
+                    rn_name,rn_ext = os.path.splitext(imgaddress)
+                    os.rename(imgaddress,"{}{}{}{}".format(rn_name,"_",str(int(time())),rn_ext))
             imgdomain = 'https://i.4pcdn.org/'
             imgurl = "{}{}/{}s.jpg".format(imgdomain,boardcode,str(post["tim"]))
             urllib.request.urlretrieve(imgurl,imgaddress)
@@ -380,192 +319,97 @@ def scrapefile(threadaddress,post,modus,boardcode,threadopno,keyword):
 
 ################################################################################
 
-def viewscraping():
-    nonemptyBoards_keywords = [b for b in configjson["boards"] if configjson["boards"][b]["keywords"]]
-    if not nonemptyBoards_keywords:
+def viewRequests():
+    someRequests = False
+    for board in cm.valueGet("downloaded"):
+        cm.denyNextAutosave()
+        requests = cm.tpt_getTasksInTier("downloaded/{}".format(board),"special")
+        if requests:
+            if not someRequests:
+                print("Current special requests:")
+                someRequests = True
+            for request in requests:
+                print("/{}/:{}:{}".format(board,str(request[0]),request[1]))
+    if not someRequests:
+        print("Currently no special requests")
+        return False
+    else:
+        return True
+
+################################################################################
+
+def viewKeywords():
+    someKeywords = False
+    for board in cm.valueGet("downloaded"):
+        cm.denyNextAutosave()
+        keywords_wl = cm.tpt_getkeywords_wl("downloaded/{}".format(board))
+        if keywords_wl:
+            if not someKeywords:
+                print("Currently scraping:")
+                someKeywords = True
+            print("/{}/:".format(board),", ".join(keywords_wl))
+    if not someKeywords:
         print("Currently not scraping any boards")
         return False
     else:
-        print("Currently scraping:")
-        for board in nonemptyBoards_keywords:
-            print("/{}/: ".format(board),end="")
-            for keyword in configjson["boards"][board]["keywords"][:-1]:
-                print(keyword,end=", ")
-            print(configjson["boards"][board]["keywords"][-1])
+        return True
 
 ################################################################################
 
-def viewrequests():
-    nonemptyBoards_requests = [b for b in configjson["boards"] if configjson["boards"][b]["requests"]]
-    if not nonemptyBoards_requests:
-        print("Currently no special requests")
-    else:
-        print("Current special requests:")
-        for board in nonemptyBoards_requests:
-            for req in configjson["boards"][board]["requests"]:
-                print("/{}/:{}:{}".format(board,str(req[0]),req[1]))
-
-################################################################################
-
-def viewblacklisting():
-    nonemptyBoards_blacklist = [b for b in configjson["boards"] if configjson["boards"][b]["blacklist"]]
-    if not nonemptyBoards_blacklist:
+def viewBlacklisting():
+    someBlacklisted = False
+    for board in cm.valueGet("downloaded"):
+        cm.denyNextAutosave()
+        idnos_bl = cm.tpt_getidnos_bl("downloaded/{}".format(board))
+        if idnos_bl:
+            if not someBlacklisted:
+                print("Currently blacklisting:")
+                someBlacklisted = True
+            print("/{}/:".format(board),", ".join([str(opno) for opno in idnos_bl]))
+    if not someBlacklisted:
         print("Currently not blacklisting any threads")
+        return False
     else:
-        print("Currently blacklisting:")
-        for board in nonemptyBoards_blacklist:
-            print("/{}/: ".format(board),end="")
-            for opno in configjson["boards"][board]["blacklist"][:-1]:
-                print(str(opno),end=", ")
-            print(str(configjson["boards"][board]["blacklist"][-1]))
+        return True
 
 ################################################################################
 
-def request(board,opno):
-    global configjson
+#Main Thread Here
+lock = threading.Lock()
+pm = saostatusmsgs.progressmsg()
 
-    try:
-        old_req = [req for req in configjson["boards"][board]["requests"] if req[0] == opno][0]
-        configjson["boards"][board]["requests"].remove(old_req)
-        old_req_old_keyword = old_req[1]
-        old_req[1] = "_OLDREQ_" + old_req_old_keyword
-        #impossible to put underscore in manually, never a keyword; former-request with board keyword that is removed from requests will not stick in active if the scraper would not naturally pick it again for that keyword
-        configjson["boards"][board]["active"].append(old_req)
-        print("Thread /{}/:{}:{} removed from special requests".format(board,str(opno),old_req_old_keyword))
-        return
-    except:
-        pass
+saotitle.printLogoTitle(title="Bateman\'s 4chan Scraper",subtitle="Version {}".format(version))
+cm = saoconfigmanager.configmanager(filename="scraperconfig.json",default={"versioncreated":version, "downloaded":{}})
+cm.tpt_manageDirectories = True
+cm.tpt_manageDirectoriesDeleteEmptyOnUpdate = True # maybe pull this ugly stuff out into either a synced
+                                                   # part of configmanager or just code on its own here
 
-    keyword = input("What keyword(s) to tag folder with? ").lower().replace("_"," ").strip()
-    if not keyword:
-        keyword = "request"
+if saovcs.olderThan(cm.valueGet("versioncreated"),"3.0.0"):
+    # update config from v2 to v3
+    for board in cm.valueGet("boards"):
+        cm.denyNextAutosave(5)
+        cm.valueMove("boards/{}/keywords".format(board),"downloaded/{}/keywords_wl".format(board))
+        cm.valueMove("boards/{}/blacklist".format(board),"downloaded/{}/idnos_bl".format(board))
+        cm.valueMove("boards/{}/doneops".format(board),"downloaded/{}/idnos_done".format(board))
+        cm.valueMove("boards/{}/active".format(board),"downloaded/{}/tiers/normal".format(board))
+        cm.valueMove("boards/{}/requests".format(board),"downloaded/{}/tiers/special".format(board))
+        if os.path.isdir(board):
+            cm.ffm_makedirs("downloaded")
+            cm.ffm_tryMove(board,"downloaded")
+        cm.valueDelete("boards")
+    cm.valueSet("versioncreated",version)
+    print("Updated config from v2 to v3")
 
-    try:
-        new_request = [t for t in configjson["boards"][board]["active"] if t[0] == opno][0]
-        configjson["boards"][board]["active"].remove(new_request)
-        try:
-            folder_old = "{}\\{} {}".format(board,str(opno),new_request[1])
-            folder_new = "{}\\{} {}".format(board,str(opno),keyword)
-            os.rename(folder_old,folder_new)
-        except:
-            pass
-        new_request[1] = keyword
-    except:
-        new_request = [opno,keyword,[]]
+#Main loop
+while True:
+    print('\n')
+    action = input("What do you want to do? (SCRAPE/SCRAPEQUIT/REQUEST/PLEBREQUEST/BLACKLIST/VIEW/ADD/DELETE/HELP/QUIT) ").upper().strip()
+    print('\n')
 
-    configjson["boards"][board]["requests"].append(new_request)
-    print("Thread /{}/:{}:{} added to special requests".format(board,str(opno),keyword))
+    if action in ["QUIT","Q"]:
+        break
 
-################################################################################
-
-def plebrequest(boardcode,keyword):
-    global boxestocheckfor,plebsHTTPHeader,configjson
-    print("Searching 4plebs archive for threads on /{}/ containing \'{}\'".format(boardcode,keyword))
-    opnos = []
-    boxno = 0
-    boxnos = len(boxestocheckfor["4plebs"])
-    for boxtocheckfor in boxestocheckfor["4plebs"]:
-        boxno += 1
-        searchjson_url = 'http://archive.4plebs.org/_/api/chan/search/?type=op&boards={}&{}={}'.format(boardcode,boxtocheckfor,keyword.replace(" ","%20"))
-        cooldown_loop = True
-        while cooldown_loop:
-            searchjson_file = urllib.request.urlopen(urllib.request.Request(searchjson_url,None,{'User-Agent':plebsHTTPHeader}))
-            searchjson = json.load(searchjson_file)
-            if "error" in searchjson:
-                if searchjson["error"] == "No results found.":
-                    cooldown_loop = False
-                elif searchjson["error"].startswith("Search limit exceeded."):
-                    try:
-                        sleeptime = 1 + int(searchjson["error"][35:37].strip()) #35 to 37 hardcoded for time
-                    except:
-                        sleeptime = 30
-                    print("Sleeping for {} seconds (4plebs cooldown {}/{})".format(str(sleeptime),str(boxno),str(boxnos)))
-                    sleep(sleeptime)
-            else:
-                for post in searchjson["0"]["posts"]:
-                    opnos.append(int(post["num"]))
-                cooldown_loop = False
-
-    opnos = list(set(opnos).difference(set([req[0] for req in configjson["boards"][boardcode]["requests"]])))
-    for opno in opnos:
-        already_active = [t for t in configjson["boards"][boardcode]["active"] if t[0] == opno]
-        if already_active:
-            new_request = already_active[0]
-            configjson["boards"][boardcode]["active"].remove(new_request)
-            new_request[1] = keyword
-        else:
-            new_request = [opno,keyword,[]]
-        configjson["boards"][boardcode]["requests"].append(new_request)
-    if opnos:
-        opnos_len = len(opnos)
-        print("Added {} special request{}".format(opnos_len,"" if opnos_len==1 else "s"))
-    else:
-        print("No more special requests added")
-
-################################################################################
-
-class class_progressmsg():
-    def __init__(self):
-        self.resetvars()
-
-    def resetvars(self):
-        self.msg = ""
-        self.pos = 0
-        self.of = 1
-        self.bsnum = 0
-        self.active = False
-
-    def progmsg(self,*args,**kwargs):
-        if 'msg' in kwargs:
-            self.msg = kwargs['msg']
-        if 'pos' in kwargs:
-            self.pos = int(kwargs['pos'])
-        if 'of' in kwargs:
-            self.of = kwargs['of']
-        if self.active == True:
-            stdout.write('\n')
-            stdout.flush()
-        stdout.write(self.msg)
-        stdout.flush()
-        self.printprog()
-        self.active = True
-
-    def printprog(self):
-        hashund = ('#'*int(10*(self.pos/self.of))).ljust(10,'_')
-        prog = '[{}] ({}/{})'.format(hashund,self.pos,self.of)
-        self.bsnum = len(prog)
-        stdout.write(prog)
-        stdout.flush()
-
-    def tick(self,times=1):
-        if self.active:
-            self.pos += times
-            self.pos = min([self.pos,self.of])
-            stdout.write('\b'*self.bsnum)
-            self.printprog()
-            if self.pos == self.of:
-                self.finish()
-
-    def finish(self):
-        if self.active:
-            stdout.write('\n')
-            stdout.flush()
-        self.resetvars()
-
-################################################################################
-
-# def gethashhex(path,blocksize=65536):
-#     with open(path,'rb') as file:
-#         hasher = md5()
-#         buffer = file.read(blocksize)
-#         while len(buffer) > 0:
-#             hasher.update(buffer)
-#             buffer = file.read(blocksize)
-#     return hasher.hexdigest()
-
-################################################################################
-
-def printhelp():
+    elif action in ["HELP","H"]:
         print("This is Bateman's 4chan scraper. It saves attachments from threads whose OPs contain a keyword of interest that is being searched for. Special requests can be made. 4plebs is also sourced")
         print("The file 'scraperconfig.json' stores the program's config in the program's directory")
         print("Scraped files are saved in nested directories in the same directory as the program\n")
@@ -581,71 +425,13 @@ def printhelp():
         print("HELP        /  H: Shows this help text")
         print("QUIT        /  Q: Closes the program")
 
-################################################################################
-
-def printTitle(title,subtitle):
-    logo = [
-    u"\u2588\u2580\u2580\u2580\u2588\u2580\u2588\u2580\u2580\u2588\u2580\u2588\u2580\u2580\u2580\u2588",
-    u"\u2588  \u2588\u2588\u2584\u2588\u2588\u2588\u2588\u2584\u2588\u2588  \u2588",
-    u"\u2588  \u2588 \u2588\u2580  \u2580\u2588 \u2588  \u2588",
-    u"\u2588\u2580\u2580\u2580\u2588      \u2588\u2580\u2580\u2580\u2588",
-    u"\u2588\u2580\u2580\u2580\u2588      \u2588\u2580\u2580\u2580\u2588",
-    u"\u2588   \u2588\u2580\u2584  \u2584\u2580\u2588   \u2588",
-    u"\u2588 \u2588\u2580\u2588\u2584\u2588\u2588\u2588\u2588\u2584\u2588\u2580\u2588 \u2588",
-    u"\u2588\u2584\u2588\u2584\u2584\u2584\u2584\u2584\u2584\u2584\u2584\u2584\u2584\u2588\u2584\u2588"]
-    plen = max(len(title),len(subtitle))
-    print(logo[0])
-    print(logo[1])
-    print(logo[2]+"  "+      "".center(plen,"~"))
-    print(logo[3]+"  "+   title.center(plen,"~"))
-    print(logo[4]+"  "+subtitle.center(plen,"~"))
-    print(logo[5]+"  "+      "".center(plen,"~"))
-    print(logo[6])
-    print(logo[7])
-    # print()
-
-################################################################################
-
-#Main Thread Here
-
-lock = threading.Lock()
-progressmsg = class_progressmsg()
-
-printTitle("BATEMAN\'S 4CHAN SCRAPER","Version {}".format(version))
-
-#Inform of auto-updating removed
-print("\nAuto-updating now removed; use git to pull latest updates")
-
-#Load or create config JSON
-if os.path.exists('scraperconfig.json'):
-    with open('scraperconfig.json') as configjson_file:
-        configjson = json.load(configjson_file)
-else:
-    configjson = new_config()
-    saveconfig()
-    print("\nCreated config file 'scraperconfig.json'")
-
-#Main loop
-while True:
-    print('\n')
-    action = input("What do you want to do? (SCRAPE/SCRAPEQUIT/REQUEST/PLEBREQUEST/BLACKLIST/VIEW/ADD/DELETE/HELP/QUIT) ").upper().strip()
-    print('\n')
-
-    if action in ["QUIT","Q"]:
-        break
-
-    elif action in ["HELP","H"]:
-        printhelp()
-
-    elif action in ["SCRAPE","S"]:
+    elif action in ["SCRAPE","S","SCRAPEQUIT","SQ"]:
         scrape()
-
-    elif action in ["SCRAPEQUIT","SQ"]:
-        scrape()
-        break
+        if action in ["SCRAPEQUIT","SQ"]:
+            break
 
     elif action in ["REQUEST","R"]:
-        viewrequests()
+        viewRequests()
         board = input("\nWhich board is the thread on? ").lower().strip()
         if not board:
             print("No board supplied")
@@ -655,30 +441,69 @@ while True:
         except:
             print("Error: Invalid number")
             continue
-        possible_new_board(board)
-        request(board,opno)
-        saveconfig()
+        cm.denyNextAutosave()
+        if cm.tpt_getTaskTierByIdno("downloaded/{}".format(board),opno) == "special":
+            keyword = cm.tpt_demoteTaskByIdno("downloaded/{}".format(board),opno)[1]
+            print("Thread /{}/:{}:{} removed from special requests".format(board,str(opno),keyword))
+        else:
+            keyword = cm.tpt_sanitiseKeyword(input("What keyword(s) to tag request with? "))
+            if not keyword:
+                keyword = "request"
+            cm.tpt_promoteTaskToByIdno("downloaded/{}".format(board),opno,keyword=keyword,promotionTier="special")
+            print("Thread /{}/:{}:{} added to special requests".format(board,str(opno),keyword))
 
     elif action in ["PLEBREQUEST","P","PR"]:
-        viewrequests()
+        viewRequests()
         board = input("\nWhat board to search on? ").lower().strip()
         if not board:
             print("No board supplied")
             continue
-        elif not board in plebboards:
+        elif not board in plebBoards:
             print("/{}/ is not a 4plebs board".format(board))
-            print("4plebs boards are: /{}/".format("/, /".join(plebboards)))
+            print("4plebs boards are: /{}/".format("/, /".join(plebBoards)))
             continue
-        keyword = input("What keyword to search 4plebs for? ").lower().replace("_"," ").strip()
+        keyword = cm.tpt_sanitiseKeyword(input("What keyword to search 4plebs for? "))
         if not keyword:
             print("No keyword supplied")
             continue
-        possible_new_board(board)
-        plebrequest(board,keyword)
-        saveconfig()
+        print("Searching 4plebs archive for threads on /{}/ containing \'{}\'".format(board,keyword))
+        opnos = []
+        boxno = 0
+        boxnos = len(boxestocheckfor["4plebs"])
+        for boxtocheckfor in boxestocheckfor["4plebs"]:
+            boxno += 1
+            searchjson_url = 'http://archive.4plebs.org/_/api/chan/search/?type=op&boards={}&{}={}'.format(board,boxtocheckfor,keyword.replace(" ","%20"))
+            cooldown_loop = True
+            while cooldown_loop:
+                searchjson_file = urllib.request.urlopen(urllib.request.Request(searchjson_url,None,{'User-Agent':plebsHTTPHeader}))
+                searchjson = json.load(searchjson_file)
+                if "error" in searchjson:
+                    if searchjson["error"] == "No results found.":
+                        cooldown_loop = False
+                    elif searchjson["error"].startswith("Search limit exceeded."):
+                        try:
+                            sleeptime = 1 + int(searchjson["error"][35:37].strip()) #35 to 37 hardcoded for time
+                        except:
+                            sleeptime = 30
+                        print("Sleeping for {} seconds (4plebs cooldown {}/{})".format(str(sleeptime),str(boxno),str(boxnos)))
+                        sleep(sleeptime)
+                else:
+                    for post in searchjson["0"]["posts"]:
+                        opnos.append(int(post["num"]))
+                    cooldown_loop = False
+
+        opnos = list(set(opnos).difference(set([req[0] for req in cm.tpt_getTasksInTier("downloaded/{}".format(board),"special")])))
+        if opnos:
+            opnos_len = len(opnos)
+            print("Added {} special request{}".format(opnos_len,"" if opnos_len==1 else "s"))
+            for opno in opnos:
+                cm.tpt_promoteTaskToByIdno("downloaded/{}".format(board),opno,keyword=keyword,promotionTier="special")
+                print("/{}/:{}:{}".format(board,str(opno),keyword))
+        else:
+            print("No more special requests added")
 
     elif action in ["BLACKLIST","B","BLACK","BL"]:
-        viewblacklisting()
+        viewBlacklisting()
         board = input("\nWhich board is the thread on? ").lower().strip()
         if not board:
             print("No board supplied")
@@ -688,76 +513,52 @@ while True:
         except:
             print("Error: Invalid number")
             continue
-        possible_new_board(board)
-        if blacklistopno in configjson["boards"][board]["blacklist"]:
-            configjson["boards"][board]["blacklist"].remove(blacklistopno)
+        if cm.tpt_idnos_blToggle("downloaded/{}".format(board),[blacklistopno])[0]:
             print("No longer blacklisting /{}/:{}".format(board,str(blacklistopno)))
         else:
-            configjson["boards"][board]["blacklist"].append(blacklistopno)
             print("Now blacklisting /{}/:{}".format(board,str(blacklistopno)))
-        saveconfig()
 
     elif action in ["VIEW","V"]:
-        viewrequests()
+        viewRequests()
         print()
-        viewscraping()
+        viewKeywords()
         print()
-        viewblacklisting()
+        viewBlacklisting()
 
     elif action in ["ADD","A"]:
-        viewscraping()
+        viewKeywords()
         board = input("\nWhich board to add keywords to? ").lower().strip()
         if not board:
             print("No board supplied")
             continue
         keywordstoadd = input("Which keywords to start scraping for? ").lower().split()
-        keywordstoadd = [keyword.replace("_"," ").strip() for keyword in keywordstoadd if keyword.replace("_"," ").strip() != ""]
-        possible_new_board(board)
-        addedcounter = 0
-        for keyword in keywordstoadd:
-            if not keyword in configjson["boards"][board]["keywords"]:
-                configjson["boards"][board]["keywords"].append(keyword)
-                addedcounter += 1
-        configjson["boards"][board]["keywords"] = sorted(configjson["boards"][board]["keywords"])
-        if addedcounter == 0:
-            if not configjson["boards"][board]["keywords"]:
+        keywordstoadd = [keyword.replace("_"," ").strip() for keyword in keywordstoadd if keyword.replace("_"," ").strip()]
+        if not cm.tpt_keywords_wlAdd("downloaded/{}".format(board),keywordstoadd):
+            if not cm.tpt_getkeywords_wl("downloaded/{}".format(board)):
                 print("No keywords added for /{}/, not scraping it".format(board))
             else:
                 print("No more keywords added for /{}/".format(board))
         else:
-            print("Keywords for /{}/ updated to: ".format(board),end="")
-            for keyword in configjson["boards"][board]["keywords"][:-1]:
-                print(keyword,end=", ")
-            print(configjson["boards"][board]["keywords"][-1])
-        saveconfig()
+            print("Keywords for /{}/ updated to:".format(board),", ".join(cm.tpt_getkeywords_wl("downloaded/{}".format(board))))
 
     elif action in ["DELETE","DEL","D"]:
-        if viewscraping() is False:
+        if viewKeywords() is False:
             continue
         board = input("\nWhich board to delete keywords from? ").lower().strip()
         if not board:
             print("No board supplied")
             continue
-        if not board in configjson["boards"] or not configjson["boards"][board]:
+        if not cm.valuePing("downloaded/{}".format(board)) or not cm.tpt_getkeywords_wl("downloaded/{}".format(board)):
             print("Currently not scraping /{}/".format(board))
             continue
         keywordstodel = input("Which keywords to stop scraping for? ").lower().split()
-        keywordstodel = [keyword.replace("_"," ").strip() for keyword in keywordstodel if keyword.replace("_"," ").strip() != ""]
-        delcounter = 0
-        for keyword in keywordstodel:
-            if keyword in configjson["boards"][board]["keywords"]:
-                configjson["boards"][board]["keywords"].remove(keyword)
-                delcounter += 1
-        if delcounter == 0:
+        keywordstodel = [keyword.replace("_"," ").strip() for keyword in keywordstodel if keyword.replace("_"," ").strip()]
+        if not cm.tpt_keywords_wlRemove("downloaded/{}".format(board),keywordstodel):
             print("No keywords removed for /{}/".format(board))
-        elif not configjson["boards"][board]["keywords"]:
+        elif not cm.tpt_getkeywords_wl("downloaded/{}".format(board)):
             print("Stopped scraping /{}/".format(board))
         else:
-            print("Keywords for /{}/ updated to: ".format(board),end="")
-            for keyword in configjson["boards"][board]["keywords"][:-1]:
-                print(keyword,end=", ")
-            print(configjson["boards"][board]["keywords"][-1])
-        saveconfig()
+            print("Keywords for /{}/ updated to:".format(board),", ".join(cm.tpt_getkeywords_wl("downloaded/{}".format(board))))
 
     else:
         print("Unknown command")